+++ tests/init_functions/main.rs
@@ -124,7 +124,7 @@ pub fn __go_init_6() {
     setup_logging();
 
         // Validate configuration
-    if (*configData.borrow().as_ref().unwrap()).len() == 0 {
+    if ((*configData.borrow().as_ref().unwrap()).len() as i32) == (0 as i32) {
         println!("{}", "Warning: No configuration data found".to_string());
     } else {
         print!("Configuration loaded with {} entries\n", (*configData.borrow().as_ref().unwrap()).len());
